+++ server/controllers/api/eventsController.js
@@ -1,32 +1,21 @@
+const eventsController = require('express').Router();
+const eventsORM = require('./eventsORM');
+
 const db = require('../../models');
 
-// Defining methods for the EventsController
-module.exports = {
-  findAll: function(req, res) {
-    db.Event.find(req.query)
-      .sort({ date: -1 })
-      .then(dbModel => res.json(dbModel))
-      .catch(err => res.status(422).json(err));
-  },
-  findById: function(req, res) {
-    db.Event.findById(req.params.id)
-      .then(dbModel => res.json(dbModel))
-      .catch(err => res.status(422).json(err));
-  },
-  create: function(req, res) {
-    db.Event.create(req.body)
-      .then(dbModel => res.json(dbModel))
-      .catch(err => res.status(422).json(err));
-  },
-  update: function(req, res) {
-    db.Event.findOneAndUpdate({ _id: req.params.id }, req.body)
-      .then(dbModel => res.json(dbModel))
-      .catch(err => res.status(422).json(err));
-  },
-  remove: function(req, res) {
-    db.Event.findById({ _id: req.params.id })
-      .then(dbModel => dbModel.remove())
-      .then(dbModel => res.json(dbModel))
-      .catch(err => res.status(422).json(err));
-  }
-};
+eventsController.post('/', (req, res) => {
+  // const { email, password } = req.body;
+  // db.events.create({ email, password })
+  //   .then(user => res.json(user))
+  //   .catch(err => res.json(err));
+});
+
+eventsController.get('/', (req, res) => {
+  eventsORM.findAll(req, res);
+});
+
+eventsController.get('/:id', (req, res) => {
+  eventsORM.findById(req, res);
+});
+
+module.exports = eventsController;
