+++ server/controllers/api/eventsORM.js
@@ -1,0 +1,37 @@
+const db = require('../../models');
+
+// Defining methods for the EventsController
+module.exports = {
+  findAll: function(req, res) {
+    console.log(req.query);
+    // res.send('test response');
+    db.Event.find(req.query)
+      .sort({ date: -1 })
+      .then(dbModel => {
+        console.log(dbModel);
+        res.json({});
+      })
+      .catch(err => res.status(422).json(err));
+  },
+  findById: function(req, res) {
+    db.Event.findById(req.params.id)
+      .then(dbModel => res.json(dbModel))
+      .catch(err => res.status(422).json(err));
+  },
+  create: function(req, res) {
+    db.Event.create(req.body)
+      .then(dbModel => res.json(dbModel))
+      .catch(err => res.status(422).json(err));
+  },
+  update: function(req, res) {
+    db.Event.findOneAndUpdate({ _id: req.params.id }, req.body)
+      .then(dbModel => res.json(dbModel))
+      .catch(err => res.status(422).json(err));
+  },
+  remove: function(req, res) {
+    db.Event.findById({ _id: req.params.id })
+      .then(dbModel => dbModel.remove())
+      .then(dbModel => res.json(dbModel))
+      .catch(err => res.status(422).json(err));
+  }
+};
